+++ src/brekekejs/index.d.ts
@@ -65,6 +65,7 @@ export type GetPalOptions = {
   tenant: string
   login_user: string
   login_password: string
+  phone_idx: string
   _wn: string
   park: string[]
   voicemail: string
@@ -91,6 +92,7 @@ export type Pbx = PbxPal & {
   notify_park?(e: PbxEvent['park']): void
   notify_voicemail?(e: PbxEvent['voicemail']): void
   notify_callrecording?(e: PbxEvent['callRecording']): void
+  notify_pal?(e: PbxEvent['pal']): void
   // not actually exist in the sdk, should be added manually
   call_pal<K extends keyof PbxPal>(
     k: K,
@@ -120,6 +122,10 @@ export type PbxEvent = {
     user: string
     talker_id: string
     status: string // on or off
+  }
+  pal: {
+    code: number
+    message: string
   }
 }