+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### 2.15.2
 
+- Implement notify_pal to show error when login the same phone index on another device
 - Implement webphone.recents.max, default 200, max 1000
 - Implement copy phone number to clipboard
 - Fix web outgoing call should not make twice (issue 985)